fix(plugin-vue): return rerender function when user empty the template
